refactor(core/statistics): unify and optimize project statistics logic
- Migrate `buildProjectStats` functionality to `Statistics` with improved structure and reusability.
- Adjust references across the codebase to use `Statistics.buildProjectStats`.
