+++ src/org/omegat/core/data/RealProject.java
@@ -75,7 +75,6 @@ import org.omegat.core.data.TMXEntry.ExternalLinked;
 import org.omegat.core.events.IProjectEventListener;
 import org.omegat.core.segmentation.SRX;
 import org.omegat.core.segmentation.Segmenter;
-import org.omegat.core.statistics.CalcStandardStatistics;
 import org.omegat.core.statistics.Statistics;
 import org.omegat.core.statistics.StatisticsInfo;
 import org.omegat.core.statistics.StatsResult;
@@ -397,7 +396,7 @@ public class RealProject implements IProject {
             loadOtherLanguages();
 
             // build word count
-            StatsResult stat = CalcStandardStatistics.buildProjectStats(this);
+            StatsResult stat = Statistics.buildProjectStats(this);
             stat.updateStatisticsInfo(hotStat);
             Statistics.writeStat(config.getProjectInternal(), stat);
 
@@ -733,7 +732,7 @@ public class RealProject implements IProject {
         // Ticket 1690 - build project statistics files
         // so that contents of these files is up to date with target files
         // sent at same moment
-        StatsResult stat = CalcStandardStatistics.buildProjectStats(this);
+        StatsResult stat = Statistics.buildProjectStats(this);
         stat.updateStatisticsInfo(hotStat);
         String fn = config.getProjectInternal() + OConsts.STATS_FILENAME;
         Statistics.writeStat(fn, stat.getTextData());
@@ -864,7 +863,7 @@ public class RealProject implements IProject {
                 LastSegmentManager.saveLastSegment();
 
                 // update statistics
-                StatsResult stat = CalcStandardStatistics.buildProjectStats(this);
+                StatsResult stat = Statistics.buildProjectStats(this);
                 stat.updateStatisticsInfo(hotStat);
                 Statistics.writeStat(config.getProjectInternal(), stat);
             } finally {
